+++ ff4j-core/src/main/java/org/ff4j/property/BasePropertyBean.java
@@ -1,0 +1,123 @@
+package org.ff4j.property;
+
+import java.util.Set;
+
+/**
+ * Created by georgekankava on 6/7/16.
+ */
+public class BasePropertyBean {
+
+    /** unique identifier for the property. */
+    protected String name;
+
+    /** property description if exist. */
+    protected String description;
+
+    /** nature of the property (classname). */
+    protected String type;
+
+    /** Value as String. */
+    protected String value;
+
+    /** Fixed values as String. */
+    protected Set< String > fixedValues = null;
+
+    public BasePropertyBean() {
+        super();
+    }
+
+    /**
+     * Getter accessor for attribute 'name'.
+     *
+     * @return
+     *       current value of 'name'
+     */
+    public String getName() {
+        return name;
+    }
+
+    /**
+     * Setter accessor for attribute 'name'.
+     * @param name
+     *      new value for 'name '
+     */
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    /**
+     * Getter accessor for attribute 'description'.
+     *
+     * @return
+     *       current value of 'description'
+     */
+    public String getDescription() {
+        return description;
+    }
+
+    /**
+     * Setter accessor for attribute 'description'.
+     * @param description
+     *      new value for 'description '
+     */
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    /**
+     * Getter accessor for attribute 'type'.
+     *
+     * @return
+     *       current value of 'type'
+     */
+    public String getType() {
+        return type;
+    }
+
+    /**
+     * Setter accessor for attribute 'type'.
+     * @param type
+     *      new value for 'type '
+     */
+    public void setType(String type) {
+        this.type = type;
+    }
+
+    /**
+     * Getter accessor for attribute 'value'.
+     *
+     * @return
+     *       current value of 'value'
+     */
+    public String getValue() {
+        return value;
+    }
+
+    /**
+     * Setter accessor for attribute 'value'.
+     * @param value
+     *      new value for 'value '
+     */
+    public void setValue(String value) {
+        this.value = value;
+    }
+
+    /**
+     * Getter accessor for attribute 'fixedValues'.
+     *
+     * @return
+     *       current value of 'fixedValues'
+     */
+    public Set<String> getFixedValues() {
+        return fixedValues;
+    }
+
+    /**
+     * Setter accessor for attribute 'fixedValues'.
+     * @param fixedValues
+     *      new value for 'fixedValues '
+     */
+    public void setFixedValues(Set<String> fixedValues) {
+        this.fixedValues = fixedValues;
+    }
+}
